requirements.txt
# Core dependencies
pandas
python-dateutil
tqdm
pyyaml
spacy

# Document processing
python-docx
PyPDF2

# Output formats
openpyxl
xlsxwriter

# Development & Testing
pytest
pytest-asyncio
black
flake8